Match paysimple styles to other pages
